backend/requirements.txt
# FastAPI and server
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
httpx>=0.27.0,<1
beautifulsoup4>=4.12.0,<5
playwright>=1.49.0,<2
google-generativeai>=0.8.0,<1
langchain-core>=0.3.0,<1
langchain-google-genai>=2.0.0,<3

# Config
pydantic>=2.0,<3
pydantic-settings>=2.0,<3
